best 500 rows
The healthcare data related to medical aid claims is collected from Kaggle and analyzed using Exploratory Data Analysis (EDA) to uncover key insights. From the dataset, the best 500 rows are identified based on data quality and relevance. These selected rows are then saved in CSV format for further use. Finally, the CSV file is uploaded into a PostgreSQL database, where the data is stored and can be easily accessed for analysis or reporting.